create atomic swap helper and update atomic swap ctrt
